refactor: Rename logging module to mfg_logging to avoid stdlib shadowing
BREAKING CHANGE: The utils.logging module has been renamed to utils.mfg_logging

Fixes circular import and AttributeError when running modules as __main__.
The old mfg_pde.utils.logging package was shadowing Python's stdlib logging
module, causing import errors when modules were executed as scripts.

Changes:
- Rename mfg_pde/utils/logging/ → mfg_pde/utils/mfg_logging/
- Update all 34 import statements across codebase
- Add backward compatibility alias with deprecation warning
- Update documentation and file path references

Migration:
- Old: from mfg_pde.utils.logging import get_logger
- New: from mfg_pde.utils.mfg_logging import get_logger
- Or:  from mfg_pde.utils import get_logger (recommended)

Backward compatibility maintained until v0.13.0 via alias.

Resolves the architectural issue preventing inline smoke tests.
