+++ examples/advanced/optimization/primal_dual_constrained_example.py
@@ -29,7 +29,7 @@ from mfg_pde.core.lagrangian_mfg_problem import (
     LagrangianComponents,
     LagrangianMFGProblem,
 )
-from mfg_pde.utils.logging import configure_research_logging, get_logger
+from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
 from mfg_pde.utils.numpy_compat import trapezoid
 
 # Configure logging
