+++ examples/applications/economics/santa_fe_bar_demo.py
@@ -40,7 +40,7 @@ import numpy as np
 
 from mfg_pde import ExampleMFGProblem
 from mfg_pde.factory import create_standard_solver
-from mfg_pde.utils.logging import configure_research_logging, get_logger
+from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
 
 # Configure logging
 configure_research_logging("santa_fe_bar_demo", level="INFO")
