+++ examples/basic/core_infrastructure/custom_hamiltonian_derivs_demo.py
@@ -35,7 +35,7 @@ import numpy as np
 from mfg_pde import ExampleMFGProblem
 from mfg_pde.core.mfg_problem import MFGComponents, MFGProblem
 from mfg_pde.factory import create_standard_solver
-from mfg_pde.utils.logging import configure_research_logging, get_logger
+from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
 
 # Configure logging
 configure_research_logging("custom_hamiltonian_derivs_demo", level="INFO")
