+++ examples/basic/solvers/acceleration_comparison.py
@@ -24,7 +24,7 @@ import matplotlib.pyplot as plt
 
 from mfg_pde import ExampleMFGProblem
 from mfg_pde.factory import create_standard_solver
-from mfg_pde.utils.logging import configure_research_logging, get_logger
+from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
 
 # Configure logging
 configure_research_logging("acceleration_comparison", level="INFO")
