+++ examples/basic/utilities/common_noise_lq_demo.py
@@ -29,7 +29,7 @@ import numpy as np
 
 from mfg_pde.alg.numerical.stochastic import CommonNoiseMFGSolver
 from mfg_pde.core.stochastic import OrnsteinUhlenbeckProcess, StochasticMFGProblem
-from mfg_pde.utils.logging import configure_research_logging, get_logger
+from mfg_pde.utils.mfg_logging import configure_research_logging, get_logger
 
 # Configure logging
 configure_research_logging("common_noise_lq_demo", level="INFO")
