+++ mfg_pde/alg/neural/operator_learning/base_operator.py
@@ -42,7 +42,7 @@ try:
 except ImportError:
     TORCH_AVAILABLE = False
 
-from mfg_pde.utils.logging.logger import get_logger
+from mfg_pde.utils.mfg_logging.logger import get_logger
 
 NDArray = np.ndarray[Any, np.dtype[Any]]
