+++ mfg_pde/alg/neural/operator_learning/operator_training.py
@@ -44,7 +44,7 @@ try:
 except ImportError:
     TORCH_AVAILABLE = False
 
-from mfg_pde.utils.logging.logger import get_logger
+from mfg_pde.utils.mfg_logging.logger import get_logger
 
 if TYPE_CHECKING:
     from collections.abc import Callable
